Move LayerKey to own file, create LayerKey.from_layer()
