+++ napari/components/experimental/chunk/__init__.py
@@ -1,6 +1,7 @@
-"""ChunkLoader and related.
+"""ChunkLoader module.
 """
 from ._config import async_config
 from ._loader import chunk_loader, synchronous_loading, wait_for_async
-from ._request import ChunkKey, ChunkRequest, LayerKey
-from ._utils import LayerRef, get_data_id
+from ._request import ChunkKey, ChunkRequest
+from ._utils import LayerRef
+from .layer_key import LayerKey
