+++ napari/layers/image/experimental/_chunked_image_loader.py
@@ -3,7 +3,7 @@
 import logging
 from typing import Optional
 
-from ....components.experimental.chunk import ChunkKey, LayerKey, get_data_id
+from ....components.experimental.chunk import ChunkKey, LayerKey
 from .._image_loader import ImageLoader
 from ._chunked_slice_data import ChunkedSliceData
 
@@ -37,11 +37,9 @@ class ChunkedImageLoader(ImageLoader):
             True if load happened synchronously.
         """
         layer = data.layer
-        layer_key = LayerKey(
-            id(layer), get_data_id(layer.data), layer._data_level, data.indices
-        )
-
+        layer_key = LayerKey.from_layer(layer, data.indices)
         key = ChunkKey(layer_key)
+
         LOGGER.debug("ChunkedImageLoader.load: %s", key)
 
         if self.current_key is not None and self.current_key == key:
